## original

```
2014-07-18 17:26:53,852 -  /opt/mhn/server/mhn/tasks/rules.py - Fetching sources from 1 sources. -- 16 sec
2014-07-18 17:27:09,810 -  /opt/mhn/server/mhn/tasks/rules.py - Bulk importing 15709 rules.      -- 1m 55 sec
2014-07-18 17:29:04,689 -  /opt/mhn/server/mhn/tasks/rules.py - Rendering rules.                 -- 1m 57sec
```
## Stop saving/rendering the references + move regex compile to top of module

```
2014-07-18 21:14:28,908 -  /opt/mhn/server/mhn/tasks/rules.py - Fetching sources from 1 sources. -- 16 sec
2014-07-18 21:14:44,324 -  /opt/mhn/server/mhn/tasks/rules.py - Bulk importing 15709 rules.      -- 58 sec
2014-07-18 21:15:42,013 -  /opt/mhn/server/mhn/tasks/rules.py - Rendering rules.                 -- 1 sec
```